+++ secfsdstools/a_utils/downloadutils.py
@@ -4,6 +4,7 @@ Download utils to download data from the SEC website.
 
 import logging
 from time import sleep
+from typing import Dict
 
 import requests
 
@@ -29,7 +30,8 @@ class UrlDownloader:
     def download_url_to_file(self, file_url: str, target_file: str,
                              expected_size: int = None,
                              max_tries: int = 6,
-                             sleep_time: int = 1):
+                             sleep_time: int = 1,
+                             headers: Dict[str, str] = None):
         """
             downloads the content auf an url and stores it into the target-file.
             retries a download several times, if it fails
@@ -44,11 +46,12 @@ class UrlDownloader:
             max_tries (int, optional, 6): maximum retries, default is 6
             sleep_time (int, optional, 1): wait time between retries,
               default is one second
+            headers (Dict[str, str], optional, None}): additional headers
 
         Returns:
             str: the written file name
         """
-        response = self.get_url_content(file_url, max_tries, sleep_time)
+        response = self.get_url_content(file_url, max_tries, sleep_time, headers=headers)
         content = response.text
 
         if expected_size is not None:
@@ -60,7 +63,8 @@ class UrlDownloader:
     def binary_download_url_to_file(self, file_url: str,
                                     target_file: str,
                                     max_tries: int = 6,
-                                    sleep_time: int = 1):
+                                    sleep_time: int = 1,
+                                    headers: Dict[str, str] = None):
         """
             downloads the binary of an url and stores it into the target-file.
             retries a download several times, if it fails
@@ -71,13 +75,15 @@ class UrlDownloader:
               (it will be written into a zipfile)
             max_tries (int, optional, 6): maximum retries, default is 6
             sleep_time (int, optional, 1): wait time between retries, default is one second
+            headers (Dict[str, str], optional, None}): additional headers
         """
-        response = self.get_url_content(file_url, max_tries, sleep_time)
+        response = self.get_url_content(file_url, max_tries, sleep_time, headers=headers)
 
         with open(target_file, "wb") as target_fp:
             target_fp.write(response.content)
 
-    def get_url_content(self, file_url: str, max_tries: int = 6, sleep_time: int = 1) \
+    def get_url_content(self, url: str, max_tries: int = 6,
+                        sleep_time: int = 1, headers: Dict[str, str] = None) \
             -> requests.models.Response:
         """
             downloads the content auf an url and returns it as a string.
@@ -85,9 +91,10 @@ class UrlDownloader:
             Uses the defined user-agent as header information
 
         Args:
-            file_url (str): url that referencese the file to be downloaded
+            url (str): url that referencese the file to be downloaded
             max_tries (int, optional, 6): maximum number of tries to get the data
             sleep_time (int, optional, 1): wait time between retries, default is one second
+            headers (Dict[str, str], optional, None}): additional headers
 
         Returns:
              requests.models.Response
@@ -97,13 +104,17 @@ class UrlDownloader:
         while current_try < max_tries:
             current_try += 1
             try:
-                response = requests.get(file_url, timeout=10,
-                                        headers={'User-Agent': self.user_agent}, stream=True)
+                if headers is None:
+                    headers = {'User-Agent': self.user_agent}
+                else:
+                    headers.update({'User-Agent': self.user_agent})
+                response = requests.get(url, timeout=10,
+                                        headers=headers, stream=True)
                 response.raise_for_status()
                 break
             except requests.exceptions.RequestException as err:
                 if current_try >= max_tries:
-                    LOGGER.info('RequestException: failed to download %s2', file_url)
+                    LOGGER.info('RequestException: failed to download %s2', url)
                     raise err
                 sleep(sleep_time)