+++ tests/a_config/test_configmgt.py
@@ -28,14 +28,14 @@ def test_environment_variable_with_file(tmp_path):
     # create the configuration at the expected location
     config_file = str(tmp_path) + '/test.cfg'
     ConfigurationManager._write_configuration(config_file,
-                                              Configuration(db_dir='blublu',
-                                                            download_dir='blublu',
-                                                            user_agent_email='email'))
+                                              Configuration(db_dir=os.path.join(tmp_path, 'blublu'),
+                                                            download_dir=os.path.join(tmp_path, 'blublu'),
+                                                            user_agent_email='user@email.com'))
 
     with patch.dict(os.environ, {SECFSDSTOOLS_ENV_VAR_NAME: config_file}, clear=True):
         configuration = ConfigurationManager.read_config_file()
         assert configuration is not None
-        assert configuration.db_dir == 'blublu'
+        assert configuration.db_dir.endswith('blublu')
 
 
 # Tests
@@ -47,13 +47,13 @@ def test_config_file_in_cwd(tmp_path, monkeypatch: pytest.MonkeyPatch):
     config_file = str(tmp_path / DEFAULT_CONFIG_FILE)
 
     ConfigurationManager._write_configuration(config_file,
-                                              Configuration(db_dir='blabla',
-                                                            download_dir='blabla',
-                                                            user_agent_email='email'))
+                                              Configuration(db_dir=os.path.join(tmp_path, 'blabla'),
+                                                            download_dir=os.path.join(tmp_path, 'blabla'),
+                                                            user_agent_email='user@email.com'))
 
     configuration = ConfigurationManager.read_config_file()
     assert configuration is not None
-    assert configuration.db_dir == 'blabla'
+    assert configuration.db_dir.endswith('blabla')
 
 
 def test_no_config_file_in_home(tmp_path):
@@ -73,12 +73,71 @@ def test_config_file_in_home(tmp_path):
     config_file = str(tmp_path / DEFAULT_CONFIG_FILE)
 
     ConfigurationManager._write_configuration(config_file,
-                                              Configuration(db_dir='bloblo',
-                                                            download_dir='bloblo',
-                                                            user_agent_email='email'))
+                                              Configuration(db_dir=os.path.join(tmp_path, 'bloblo'),
+                                                            download_dir=os.path.join(tmp_path, 'bloblo'),
+                                                            user_agent_email='user@email.com'))
     with patch('os.path.expanduser') as mock_expanduser:
         mock_expanduser.return_value = str(tmp_path)
 
         configuration = ConfigurationManager.read_config_file()
         assert configuration is not None
-        assert configuration.db_dir == 'bloblo'
+        assert configuration.db_dir.endswith('bloblo')
+
+
+def test_check_basic_configuration(tmp_path):
+    invalid_email_config = Configuration(db_dir=str(tmp_path),
+                                         download_dir=str(tmp_path),
+                                         user_agent_email='email.com')
+
+    results = ConfigurationManager.check_basic_configuration(invalid_email_config)
+    assert len(results) == 1
+    assert 'UserAgentEmail' in results[0]
+
+
+def test_check_rapid_configuration(tmp_path):
+    invalid_rapid_plan = Configuration(db_dir=str(tmp_path),
+                                       download_dir=str(tmp_path),
+                                       user_agent_email='abc@xy.org',
+                                       rapid_api_plan='bl')
+
+    results = ConfigurationManager.check_rapid_configuration(invalid_rapid_plan)
+    assert len(results) == 1
+    assert 'RapidApiPlan' in results[0]
+
+    valid_rapid_plan = Configuration(db_dir=str(tmp_path),
+                                     download_dir=str(tmp_path),
+                                     user_agent_email='abc@xy.org',
+                                     rapid_api_plan='basic')
+
+    results = ConfigurationManager.check_rapid_configuration(valid_rapid_plan)
+    assert len(results) == 0
+
+    valid_rapid_plan = Configuration(db_dir=str(tmp_path),
+                                     download_dir=str(tmp_path),
+                                     user_agent_email='abc@xy.org',
+                                     rapid_api_plan='premium')
+    results = ConfigurationManager.check_rapid_configuration(valid_rapid_plan)
+    assert len(results) == 0
+
+    valid_rapid_plan = Configuration(db_dir=str(tmp_path),
+                                     download_dir=str(tmp_path),
+                                     user_agent_email='abc@xy.org',
+                                     rapid_api_plan=None)
+    results = ConfigurationManager.check_rapid_configuration(valid_rapid_plan)
+    assert len(results) == 0
+
+    rapid_api_key = os.environ.get('RAPID_API_KEY')
+    valid_api_key = Configuration(db_dir=str(tmp_path),
+                                  download_dir=str(tmp_path),
+                                  user_agent_email='abc@xy.com',
+                                  rapid_api_key=rapid_api_key)
+    results = ConfigurationManager.check_rapid_configuration(valid_api_key)
+    assert len(results) == 0
+
+    invalid_api_key = Configuration(db_dir=str(tmp_path),
+                                    download_dir=str(tmp_path),
+                                    user_agent_email='abc@xy.com',
+                                    rapid_api_key="abc")
+    results = ConfigurationManager.check_rapid_configuration(invalid_api_key)
+    assert len(results) == 1
+    assert 'RapidApiKey' in results[0]